continue the migration changes from Models to Repository adaptation of linked controllers

+++ src/Controller/AdminController.php
@@ -27,10 +27,16 @@ class AdminController extends CoreController
     public function dashboard()
     {
         if (!$this->userIsConnected()) {
-            $error403 = new ErrorController;
-            $error403->accessDenied();
+            $this->flashes('warning', 'Merci de te connecter!');
+            header('Location: /security/login');
         } else {
-            $this->show('admin/dashboard');
+            $currentUserRole = $this->roleRepository->findById($this->userIsConnected()->getRoleId());
+            if ($currentUserRole->getName() === "utilisateur")  {
+                $error403 = new ErrorController;
+                $error403->accessDenied();
+            } else {
+                $this->show('admin/dashboard');
+            }
         }
     }
 
@@ -43,8 +49,8 @@ class AdminController extends CoreController
     public function comments()
     {
         if (!$this->userIsConnected()) {
-            $error403 = new ErrorController;
-            $error403->accessDenied();
+            $this->flashes('warning', 'Merci de te connecter!');
+            header('Location: /security/login');
         } else {
             $comments = $this->commentRepository->findAll();
             // Stocker le user en session
@@ -52,12 +58,12 @@ class AdminController extends CoreController
             foreach ($comments as $comment) {
                 $commentId = $comment->getId();
             }
-    
+
             // Récupérer le role du user en session
             $roleId = $userCurrent->getRoleId();
             $role = $this->roleRepository->findById($roleId);
 
-            if ($role->getName() == "utilisateur") {
+            if ($role->getName() === "utilisateur") {
                 $error403 = new ErrorController;
                 $error403->accessDenied();
             } else {
@@ -65,7 +71,7 @@ class AdminController extends CoreController
                     $this->update($commentId);
                 }
             }
-            
+
             // On affiche notre vue en transmettant les infos du Comment et des messages d'alerte
             $this->show('/admin/comment/read', [
                 'comments' => $comments
@@ -81,21 +87,30 @@ class AdminController extends CoreController
     public function update(int $commentId): void
     {
         $comment = $this->commentRepository->findById($commentId);
-        
+
         if (!$this->userIsConnected()) {
-            $error403 = new ErrorController;
-            $error403->accessDenied();
+            $this->flashes('warning', 'Merci de te connecter!');
+            header('Location: /security/login');
         } else {
-            if ($this->isPost()) {
-                $status = (int)filter_input(INPUT_POST, 'status');
-                $comment->setStatus($status);
-    
-                if ($this->commentRepository->update($comment)) {
-                    $this->flashes('success', "Le Bubbles Comment $commentId a bien été modifié.");
-                    header('Location: /admin/comments');
-                    return;
-                } else {
-                    $this->flashes('danger', "Le Bubbles Comment  $commentId n'a pas été modifié!");
+            $userCurrent = $this->userIsConnected();
+            // Récupérer le role du user en session
+            $roleId = $userCurrent->getRoleId();
+            $role = $this->roleRepository->findById($roleId);
+            if ($role->getName() === "utilisateur") {
+                $error403 = new ErrorController;
+                $error403->accessDenied();
+            } else {
+                if ($this->isPost()) {
+                    $status = (int)filter_input(INPUT_POST, 'status');
+                    $comment->setStatus($status);
+
+                    if ($this->commentRepository->update($comment)) {
+                        $this->flashes('success', "Le Bubbles Comment $commentId a bien été modifié.");
+                        header('Location: /admin/comments');
+                        return;
+                    } else {
+                        $this->flashes('danger', "Le Bubbles Comment  $commentId n'a pas été modifié!");
+                    }
                 }
             }
         }
@@ -110,24 +125,30 @@ class AdminController extends CoreController
     public function delete(int $commentId): void
     {
         $comment = $this->commentRepository->findById($commentId);
-       
 
         if (!$this->userIsConnected()) {
-            $error403 = new ErrorController;
-            $error403->accessDenied();
-        } elseif ($this->userIsConnected()) {
-            $currentUserRole = $this->roleRepository->findById($this->userIsConnected()->getRoleId());
-            if ($currentUserRole->getName() === "utilisateur")
-            $error403 = new ErrorController;
-            $error403->accessDenied();
+            $this->flashes('warning', 'Merci de te connecter!');
+            header('Location: /security/login');
         } else {
-            if ($comment) {
-                $this->commentRepository->delete($commentId);
-                $this->flashes('success', "Le Bubbles Comment $commentId a bien été supprimé.");
-                header('Location: /admin/comments');
-                return;
+            if (!$this->commentRepository->findById($commentId)) {
+                $error404 = new ErrorController();
+                $error404->pageNotFoundAction();
             } else {
-                $this->flashes('danger', "Ce Bubbles Comment n'existe pas!");
+                $currentUserRole = $this->roleRepository->findById($this->userIsConnected()->getRoleId());
+                if ($currentUserRole->getName() === "utilisateur") {
+                    $error403 = new ErrorController;
+                    $error403->accessDenied();
+            } else {
+                    if ($comment) {
+                        $this->commentRepository->delete($commentId);
+                        $this->flashes('success', "Le Bubbles Comment $commentId a bien été supprimé.");
+                        header('Location: /admin/comments');
+                        return;
+                    } else {
+                        $error404 = new ErrorController();
+                        $error404->pageNotFoundAction();
+                    }
+                }
             }
         }
     }
